Throw when a migration data symbol cannot be resolved
When requiring-resolve fails to resolve the symbol it will either throw an
exception (if the namespace does not exist) or return nil (if the namespace
exists but the var does not exist).  In both cases the migration must fail.

Previously, if `requiring-resolve` returned nil the migration became a no-op
and was not performed or recorded.